split batchers into 2 structs
